requirements.txt
aiosmtplib>=3.0.2
aiosqlite>=0.20.0
APScheduler>=3.10.4
argon2-cffi>=23.1.0
asyncpg>=0.30.0
email_validator>=2.2.0
fastapi>=0.115.6
greenlet>=3.3.0
Jinja2>=3.1.4
pydantic>=2.10.3
pydantic-settings>=2.6.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.18
SQLAlchemy>=2.0.36
uvicorn[standard]>=0.34.0
